Rename cached_bls to bls_cache
